First attempt at making plotting libraries optional for OVERT
