+++ Project.toml
@@ -6,33 +6,29 @@ version = "0.1.2"
 [deps]
 Calculus = "49dc2e85-a5d0-5ad3-a950-438e2897f1b9"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
-LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-NBInclude = "0db19996-df87-5ea3-a455-e3a50d440464"
 NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
-PGFPlots = "3b7a836e-365b-5785-a47d-02c71176b4aa"
-PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 SymEngine = "123dc426-2d89-5057-bbad-38513e3affd8"
 
-[compat]
-Calculus = "0.5"
-Interpolations = "0.14"
-LaTeXStrings = "1.3"
-MacroTools = "0.5"
-NBInclude = "2.3"
-NLsolve = "4.5"
-PGFPlots = "3.4"
-PlotlyBase = "0.8"
-Plots = "1.19, 1.20, 1.38"
-Roots = "2.0"
-SymEngine = "0.8, 0.9"
-julia = "1.6"
+# Compatibility with Julia < 1.9 -- newer Julia version ignore these deps
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+PGFPlots = "3b7a836e-365b-5785-a47d-02c71176b4aa"
+PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[weakdeps]
+LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
+PGFPlots = "3b7a836e-365b-5785-a47d-02c71176b4aa"
+PlotlyBase = "a03496cd-edff-5a9b-9e67-9cda94a718b5"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+
+[extensions]
+PlotExt = ["LaTeXStrings", "PGFPlots", "Plots", "PlotlyBase"]
 
 [extras]
 NBInclude = "0db19996-df87-5ea3-a455-e3a50d440464"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "NBInclude"]
+test = ["Test", "NBInclude", "LaTeXStrings", "PGFPlots", "Plots", "PlotlyBase"]
